heron-nvila/pyproject.toml
[project]
name = "heron-nvila"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "einops>=0.8.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "s2wrapper",
    "torchvision>=0.23.0",
    "transformers==4.45.0",
]

[tool.uv.sources]
s2wrapper = { git = "https://github.com/bfshi/scaling_on_scales.git" }
